backend/requirements.txt.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
spacy==3.7.2
Faker==20.1.0
PyMuPDF==1.23.8
Pillow==10.1.0
pytesseract==0.3.10
easyocr==1.7.1
pydantic==2.5.2
python-jose==3.3.0
aiofiles==23.2.1
